requirements.txt
customtkinter~=5.1.3
Pillow